combined the function
